+++ src/core/annotation.js
@@ -26,7 +26,6 @@ import {
   BBOX_INIT,
   F32_BBOX_INIT,
   FeatureTest,
-  getModificationDate,
   info,
   isArrayEqual,
   LINE_DESCENT_FACTOR,
@@ -43,6 +42,7 @@ import {
   collectActions,
   escapeString,
   getInheritableProperty,
+  getModificationDate,
   getParentToUpdate,
   getRotationMatrix,
   IDENTITY_MATRIX,
