+++ src/core/editor/pdf_editor.js
@@ -23,17 +23,14 @@
 import {
   deepCompare,
   getInheritableProperty,
+  getModificationDate,
   getNewAnnotationsMap,
   stringToAsciiOrUTF16BE,
 } from "../core_utils.js";
 import { Dict, isName, Name, Ref, RefSet, RefSetCache } from "../primitives.js";
-import {
-  getModificationDate,
-  stringToBytes,
-  stringToPDFString,
-} from "../../shared/util.js";
 import { incrementalUpdate, writeValue } from "../writer.js";
 import { NameTree, NumberTree } from "../name_number_tree.js";
+import { stringToBytes, stringToPDFString } from "../../shared/util.js";
 import { AnnotationFactory } from "../annotation.js";
 import { BaseStream } from "../base_stream.js";
 import { StringStream } from "../stream.js";
